Move helper functions in helpers.go
